create state_validity.cpp
